Add Closure Compiler v20160315
This is done in preparation for updating the Closure Compiler version
in GWT.

Closure Compiler now provides an unshaded jar without bundled
dependencies on Maven Central. This makes the JarJar rules simpler.

Once the update is done in GWT, I will remove the other obsolete
versions of Closure Compiler from this repository.